Fixing client output.

+++ utils.c
@@ -114,7 +114,7 @@ double timespec_diff(struct timespec *start, struct timespec *end) {
         diff.tv_nsec = end->tv_nsec - start->tv_nsec;
     }
 
-    result = diff.tv_sec + diff.tv_nsec/1000000000;
+    result = diff.tv_sec + (double)diff.tv_nsec/1000000000;
     return result;
 }
